refactor(trusty-mpm): split bug_report/github.rs HTTP transport under 500-line cap (closes #496)
Move RealGithubClient + reqwest transport into github_client.rs and extract
inline tests into github_tests.rs (included via #[path]), dropping github.rs
from 805 lines to 276. No behavior change; public surface and re-exports
are unchanged.
